revised index.js file, added the express-list-endpoints module, and started testing out how I can separate route files
