+++ package.json
@@ -55,6 +55,7 @@
     "date-fns": "^1.29.0",
     "dotenv": "^5.0.1",
     "express": "^4.16.2",
+    "express-list-endpoints": "^3.0.1",
     "knex": "^0.14.4",
     "lodash": "^4.17.5",
     "pg": "^7.4.1",
