+++ src/services/users/routes/test.js
@@ -1,0 +1,15 @@
+'use strict';
+
+
+const express = require('express'),
+	app = express(),
+	router = express.Router();
+
+
+router.get('/user', (req, res) => {
+	res.status(200).send({
+		message: 'Welcome to the beginning of nothingness.',
+	});
+});
+app.use('/', router);
+module.exports = router;
